requirements.txt
# FinNutriAgent — Python Dependencies
# Tested with Python 3.10, 3.11, 3.12

# Core data
pandas>=2.0.0
numpy>=1.24.0

# Optimization
pulp>=2.7.0

# Visualization (notebook only)
matplotlib>=3.7.0
seaborn>=0.12.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.0.0

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
